# Conflicts:
#	Dairiten/Data/ApplicationDbContext.cs
#	Dairiten/Data/Migrations/20220922024216_Initial.Designer.cs
#	Dairiten/Data/Migrations/20220922024216_Initial.cs

+++ Dairiten/Models/m_tax.cs
@@ -1,5 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel;
+using Microsoft.EntityFrameworkCore;
 
 namespace Dairiten.Models
 {
@@ -9,11 +10,13 @@ namespace Dairiten.Models
         [DisplayName("開始日")]
         [DataType(DataType.Date)]
         public DateTime kaisibi { get; set; }
+
         [DisplayName("終了日")]
         [DataType(DataType.Date)]
         public DateTime shuryobi { get; set; }
+
         [DisplayName("消費税")]
         [RegularExpression(@"^\d+(\.\d{1,2})?$", ErrorMessage = "少数以下２桁以上は入力できません")]
-        public decimal tax { get; set; }
+        public float tax { get; set; }
     }
 }
